Add set -e so that the upload/download process fast-fails
We don't have a recovery mechanism, this ensures that the error is
visible instead of allowing for the possibility of some downloads or
uploads failing silently.

A future improvement would be to report a list of failed uploads or
failed downloads separately to make it easy to retry.

Arvados-DCO-1.1-Signed-off-by: Peter Amstutz <peter.amstutz@curii.com>